refactor: share DTOs between API and client
