+++ MeterReadingsApi/MeterReadingsApi.IntegrationTests/AccountsControllerIntegrationTests.cs
@@ -1,4 +1,4 @@
-﻿using MeterReadingsApi.Models;
+﻿using MeterReadingsApi.Shared;
 using System.Diagnostics.CodeAnalysis;
 using System.Text.Json;
 using Xunit;
