+++ MeterReadingsApi/MeterReadingsApi.IntegrationTests/MeterReadingsApi.IntegrationTests.csproj
@@ -22,5 +22,6 @@
   <ItemGroup>
     <ProjectReference Include="..\MeterReadingsApi\MeterReadingsApi.csproj" />
     <ProjectReference Include="..\MeterReadingsApi.DataModel\MeterReadingsApi.DataModel.csproj" />
+    <ProjectReference Include="..\MeterReadingsApi.Shared\MeterReadingsApi.Shared.csproj" />
   </ItemGroup>
 </Project>
