+++ MeterReadingsApi/MeterReadingsApi.IntegrationTests/MeterReadingsControllerIntegrationTests.cs
@@ -1,4 +1,5 @@
-﻿using MeterReadingsApi.Models;
+using MeterReadingsApi.Models;
+using MeterReadingsApi.Shared;
 using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Net.Http.Headers;
