+++ MeterReadingsApi/MeterReadingsApi.Shared/AccountDto.cs
@@ -1,0 +1,8 @@
+namespace MeterReadingsApi.Shared;
+
+public class AccountDto
+{
+    public int AccountId { get; set; }
+    public string FirstName { get; set; } = string.Empty;
+    public string LastName { get; set; } = string.Empty;
+}
