+++ MeterReadingsApi/MeterReadingsApi.Shared/MeterReadingDto.cs
@@ -1,0 +1,8 @@
+namespace MeterReadingsApi.Shared;
+
+public class MeterReadingDto
+{
+    public int AccountId { get; set; }
+    public DateTime MeterReadingDateTime { get; set; }
+    public int MeterReadValue { get; set; }
+}
