+++ MeterReadingsApi/MeterReadingsApi.UnitTests/AccountsControllerTests.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using MeterReadingsApi.Controllers;
 using MeterReadingsApi.DataModel;
-using MeterReadingsApi.Models;
+using MeterReadingsApi.Shared;
 using MeterReadingsApi.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using Moq;
