+++ MeterReadingsApi/MeterReadingsApi.sln
@@ -13,6 +13,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeterReadingsApi.Integratio
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeterReadingsBlazorClient", "MeterReadingsBlazorClient\MeterReadingsBlazorClient.csproj", "{BD58CEFD-DEF8-4F48-90DC-D0B7EAFC2B9E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeterReadingsApi.Shared", "MeterReadingsApi.Shared\MeterReadingsApi.Shared.csproj", "{B75C1C24-B465-4F5F-B631-26F5D31DFD42}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -83,6 +85,18 @@ Global
 		{BD58CEFD-DEF8-4F48-90DC-D0B7EAFC2B9E}.Release|x64.Build.0 = Release|Any CPU
 		{BD58CEFD-DEF8-4F48-90DC-D0B7EAFC2B9E}.Release|x86.ActiveCfg = Release|Any CPU
 		{BD58CEFD-DEF8-4F48-90DC-D0B7EAFC2B9E}.Release|x86.Build.0 = Release|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Debug|x64.Build.0 = Debug|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Debug|x86.Build.0 = Debug|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Release|x64.ActiveCfg = Release|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Release|x64.Build.0 = Release|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Release|x86.ActiveCfg = Release|Any CPU
+		{B75C1C24-B465-4F5F-B631-26F5D31DFD42}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
