--- MeterReadingsApi/MeterReadingsApi/Models/AccountDto.cs
@@ -1,9 +1,0 @@
-﻿namespace MeterReadingsApi.Models
-{
-    public class AccountDto
-    {
-        public int AccountId { get; set; }
-        public string FirstName { get; set; } = string.Empty;
-        public string LastName { get; set; } = string.Empty;
-    }
-}
