--- MeterReadingsApi/MeterReadingsApi/Models/MeterReadingDto.cs
@@ -1,9 +1,0 @@
-﻿namespace MeterReadingsApi.Models
-{
-    public class MeterReadingDto
-    {
-        public int AccountId { get; set; }
-        public DateTime MeterReadingDateTime { get; set; }
-        public int MeterReadValue { get; set; }
-    }
-}
