--- MeterReadingsApi/MeterReadingsBlazorClient/Models/MeterReadingDto.cs
@@ -1,9 +1,0 @@
-namespace MeterReadingsBlazorClient.Models
-{
-    public class MeterReadingDto
-    {
-        public int AccountId { get; set; }
-        public DateTime MeterReadingDateTime { get; set; }
-        public int MeterReadValue { get; set; }
-    }
-}
