+++ MeterReadingsApi/MeterReadingsBlazorClient/_Imports.razor
@@ -8,6 +8,6 @@
 @using Microsoft.JSInterop
 @using MeterReadingsBlazorClient
 @using MeterReadingsBlazorClient.Layout
-@using MeterReadingsBlazorClient.Models
+@using MeterReadingsApi.Shared
 @using Blazorise
 @using Blazorise.DataGrid
